% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{California cities}
\format{
An sf object with 1522 features and 17 fields
}
\usage{
cities
}
\description{
MULTIPOLYGON object of California cities
}
\keyword{datasets}
