% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{California counties}
\format{
An sf object with 58 features and 18 fields
}
\usage{
counties
}
\description{
MULTIPOLYGON object of California counties
}
\keyword{datasets}
