% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{California regions}
\format{
An sf object with 5 features and 2 fields
}
\usage{
regions
}
\description{
MULTIPOLYGON/POLYGON object of six Caifornia regions
}
\keyword{datasets}
