% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schools}
\alias{schools}
\title{California schools}
\format{
An sf object with 10043 features and 83 fields
}
\usage{
schools
}
\description{
POINT object of California schools
}
\keyword{datasets}
